--- web/components/KakaoMap/DummyGPSData.tsx
@@ -1,17 +1,0 @@
-export const DummyGPSData = [
-  { lat: 37.5665, lng: 126.978 },
-  { lat: 37.5666, lng: 126.9782 },
-  { lat: 37.5667, lng: 126.9784 },
-  { lat: 37.5668, lng: 126.9786 },
-  { lat: 37.5669, lng: 126.9788 },
-  { lat: 37.567, lng: 126.979 },
-  { lat: 37.5671, lng: 126.9788 },
-  { lat: 37.5673, lng: 126.9784 },
-  { lat: 37.5674, lng: 126.9782 },
-  { lat: 37.5675, lng: 126.978 },
-
-  { lat: 37.5671, lng: 126.9772 },
-  { lat: 37.567, lng: 126.977 },
-  { lat: 37.5669, lng: 126.9768 },
-  { lat: 37.5668, lng: 126.9766 }
-];
