--- web/components/personal-map/ControlButton.tsx
@@ -1,19 +1,0 @@
-// components/ui/ControlButton.tsx
-import { ReactNode } from 'react';
-
-export default function ControlButton({
-  children,
-  onClick
-}: {
-  children: ReactNode;
-  onClick?: () => void;
-}) {
-  return (
-    <button
-      onClick={onClick}
-      className="flex h-[78px] w-[78px] items-center justify-center rounded-full bg-primary"
-    >
-      {children}
-    </button>
-  );
-}
