+++ web/package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@vis.gl/react-google-maps": "^1.5.5",
     "lucide-react": "^0.536.0",
     "next": "15.4.4",
     "prettier": "^3.6.2",
